extracted code from Title into a new LightBanner component
